moved CPD.py to Factor; changed import statement for CPD in BayesianModel
