+++ pgmpy/Factor/CPD.py
